Split protocol compatibility module into focused components
