+++ crates/protocol/src/lib.rs
@@ -13,7 +13,10 @@ mod negotiation;
 mod varint;
 mod version;
 
-pub use compatibility::{CompatibilityFlags, KnownCompatibilityFlag};
+pub use compatibility::{
+    CompatibilityFlags, KnownCompatibilityFlag, KnownCompatibilityFlagsIter,
+    ParseKnownCompatibilityFlagError,
+};
 pub use envelope::{
     EnvelopeError, HEADER_LEN as MESSAGE_HEADER_LEN, LogCode, LogCodeConversionError,
     MAX_PAYLOAD_LENGTH, MPLEX_BASE, MessageCode, MessageHeader, ParseLogCodeError,
